LayoutTests/imported/w3c/web-platform-tests/css/css-grid/masonry/tentative/items/row-minimum-contribution-with-percentages-expected.txt


PASS auto - rows
FAIL min - rows assert_equals: grid-template-rows expected "50px" but got "150px"
PASS max - rows
FAIL size - rows assert_equals: grid-template-rows expected "100px" but got "110px"
FAIL min max - rows assert_equals: grid-template-rows expected "50px" but got "160px"
FAIL min size - rows assert_equals: grid-template-rows expected "50px" but got "210px"
FAIL max size - rows assert_equals: grid-template-rows expected "100px" but got "160px"
FAIL min max size - rows assert_equals: grid-template-rows expected "50px" but got "210px"
